(OPS-16955) Update MOTD
This was the MOTD chosen by the SRE MOTD contest.

It shows services defined on the host, and some metadata. It also
provides for a way to display key value information, such as the name of
the Jenkins master a worker is attached to, the current PE build.

This also moves the logic to exclude Windows into
`profile::motd::fragment` so that it can be used without condition on
all hosts.

Tested on:
  * `ldap3-test.ops.puppetlabs.net`
  * `opsrepo-aptly-dev-1.ops.puppetlabs.net`
  * `opsrepo-web-dev-1.ops.puppetlabs.net`
  * `pe-compiler6-test.ops.puppetlabs.net`
  * `pe-mom2-test.ops.puppetlabs.net`
  * `repo-ips1-prod.ops.puppetlabs.net`
  * `weth-dev.delivery.puppetlabs.net`